moved some example code around
