+++ examples/typescript/core/src/accounts/controller.ts
