+++ examples/typescript/core/src/accounts/index.ts
