+++ examples/typescript/core/src/accounts/types.ts
