+++ examples/typescript/core/src/app/controller.ts
