+++ examples/typescript/core/src/app/index.ts
