+++ examples/typescript/core/src/app/types.ts
