+++ examples/typescript/core/src/channels/index.ts
