+++ examples/typescript/core/src/core.ts
