+++ examples/typescript/core/src/ui/controller.ts
