+++ examples/typescript/core/src/ui/types.ts
